Itroducing fdSize and hSize functions
Original size was slightly modified and is now named fdSize. It's
parametrized by file descriptor instead of being fixed on stdout.

New function with the name size is also present, but it just calls
fdSize with STDOUT as an argument. This preserves backward
compatibility.

Function hSize is a wrapper around fdSize that takes Handle instead
of file descriptor.